Random things
reworked some command for economy,, worked on commands to set a welcome messsage and goodbye message (But not working)

moved shopitems to the "data" folder, and two exports for doing the roles certain commands need Admin roles (with variation names), along with mod (with varition names). made a botinfo command (very basic) and made a command to set the bots prefix.